requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-google-genai>=1.0.0
langgraph>=0.1.0
haystack-ai>=2.0.0
pydantic>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
openai>=1.0.0
anthropic>=0.8.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
typing-extensions>=4.5.0

# PDF processing libraries
PyPDF2>=3.0.0
pdfplumber>=0.9.0
PyMuPDF>=1.23.0
reportlab>=4.0.0

# Visualization libraries
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=2.0.0
numpy>=1.24.0
